Adding Product objects to store sceapped data
Creating java objects that will store the scrapped information for each
of the products that is found on the web page.

There is two types of products, a generic product and a food product.
What sets the food product apart from a generic product is the ability
to find nutritional information. This does not mean a product is not
food however as the web page may not contain the nutirtional
information.

Added tests to make sure the correct items are stored in the correct
fields on the products.